pyproject.toml
[tool.poetry]
name = "home-image-generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
starlette = "^0.32.0.post1"
uvicorn = "^0.24.0.post1"
pillow = "^10.1.0"
numpy = "^1.26.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
